# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "KKN Candles API",
        "description": "API pour la boutique de bougies KKN Candles",
        "contact": {
            "name": "Support KKN Candles",
            "email": "support@kkncandles.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Serveur API"
        }
    ],
    "paths": {
...
        "/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Liste des commandes",
                "description": "Récupérer les commandes de l'utilisateur connecté",
                "operationId": "getOrders",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Nombre de commandes par page (max 50)",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des commandes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "items": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Order"
                                                    }
                                                },
                                                "meta": {
                                                    "$ref": "#/components/schemas/PaginationMeta"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Créer une commande",
                "description": "Créer une nouvelle commande",
                "operationId": "createOrder",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email",
                                    "phone",
                                    "address",
                                    "city",
                                    "items"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "Jean"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "Dupont"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "jean@example.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "maxLength": 20,
                                        "example": "+225 07 00 00 00 00"
                                    },
                                    "address": {
                                        "type": "string",
                                        "maxLength": 500,
                                        "example": "Cocody, Rue des Jardins"
                                    },
                                    "city": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "Abidjan"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "maxLength": 1000,
                                        "example": "Livrer le matin",
                                        "nullable": true
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "product_id",
                                                "quantity"
                                            ],
                                            "properties": {
                                                "product_id": {
                                                    "type": "string",
                                                    "format": "uuid"
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "maximum": 100,
                                                    "minimum": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Commande créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Commande créée avec succès"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Order"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur lors de la création de la commande"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/orders/{order}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Détail d'une commande",
                "description": "Récupérer les détails d'une commande",
                "operationId": "getOrder",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "description": "ID de la commande",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détail de la commande",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Order"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "404": {
                        "description": "Commande non trouvée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Order": {
                "title": "Order",
                "description": "Modèle commande",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "order_number": {
                        "type": "string",
                        "example": "KKN-20240101-0001"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "confirmed",
                            "processing",
                            "shipped",
                            "delivered",
                            "cancelled"
                        ]
                    },
                    "status_label": {
                        "type": "string",
                        "example": "En attente"
                    },
                    "customer": {
                        "properties": {
                            "first_name": {
                                "type": "string"
                            },
                            "last_name": {
                                "type": "string"
                            },
                            "full_name": {
                                "type": "string"
                            },
                            "email": {
                                "type": "string",
                                "format": "email"
                            },
                            "phone": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "shipping": {
                        "properties": {
                            "address": {
                                "type": "string"
                            },
                            "city": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "subtotal": {
                        "type": "integer",
                        "example": 15000
                    },
                    "shipping_cost": {
                        "type": "integer",
                        "example": 2000
                    },
                    "total": {
                        "type": "integer",
                        "example": 17000
                    },
                    "formatted_total": {
                        "type": "string",
                        "example": "17 000 FCFA"
                    },
                    "notes": {
                        "type": "string",
                        "nullable": true
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderItem"
                        }
                    },
                    "payment": {
                        "$ref": "#/components/schemas/Payment"
                    },
                    "paid_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "shipped_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "delivered_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "OrderItem": {
                "title": "OrderItem",
                "description": "Article de commande",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "product_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "product_name": {
                        "type": "string"
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 2
                    },
                    "unit_price": {
                        "type": "integer",
                        "example": 5000
                    },
                    "subtotal": {
                        "type": "integer",
                        "example": 10000
                    }
                },
                "type": "object"
            },
            "Payment": {
                "title": "Payment",
                "description": "Modèle paiement",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "transaction_id": {
                        "type": "string"
                    },
                    "amount": {
                        "type": "integer",
                        "example": 17000
                    },
                    "currency": {
                        "type": "string",
                        "example": "XOF"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "completed",
                            "failed",
                            "cancelled"
                        ]
                    },
                    "payment_method": {
                        "type": "string",
                        "example": "cinetpay"
                    },
                    "operator": {
                        "type": "string",
                        "example": "ORANGE",
                        "nullable": true
                    },
                    "completed_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "PaginationMeta": {
                "title": "PaginationMeta",
                "description": "Métadonnées de pagination",
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 10
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 12
                    },
                    "total": {
                        "type": "integer",
                        "example": 120
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Authentification JWT Bearer Token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}